Feature/display content for users

+++ pages/api/flashcards/[id].js
@@ -1,12 +1,18 @@
 import Flashcard from "@/db/models/flashcard";
 import dbConnect from "@/db/connect";
+import { getServerSession } from "next-auth/next";
+import { authOptions } from "../auth/[...nextauth]";
 
 export default async function handler(request, response) {
+  const session = await getServerSession(request, response, authOptions);
   await dbConnect();
   const { id } = request.query;
 
   if (request.method === "PUT") {
     try {
+      if (!session) {
+        return response.status(401).json({ status: "Not authorized" });
+      }
       const flashcardData = request.body;
       const flashcardToUpdate = await Flashcard.findByIdAndUpdate(
         id,
@@ -29,6 +35,9 @@ export default async function handler(request, response) {
 
   if (request.method === "DELETE") {
     try {
+      if (!session) {
+        return response.status(401).json({ status: "Not authorized" });
+      }
       const deleted = await Flashcard.findByIdAndDelete(id);
 
       if (!deleted) {
